Moved mocks to commns class library. Also im loosing my mind because of this bug
