+++ CodeBase/Backend/Controllers/TransactionsController.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Backend.Interfaces;
 using Commons.Domain;
 using Microsoft.AspNetCore.Http;
