+++ CodeBase/Backend/Startup.cs
@@ -20,7 +20,16 @@ namespace Backend
     {
         public Startup(IConfiguration configuration)
         {
-            Configuration = configuration;
+            try
+            {
+                Configuration = configuration;
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e.Message);
+                Console.WriteLine(e.StackTrace);
+            }
+          
         }
 
         public IConfiguration Configuration { get; }
@@ -28,34 +37,52 @@ namespace Backend
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddControllers();
+            try
+            {
+                services.AddControllers();
 
-            //Add MockedServiced if in Debug, real services if not
+                //Add MockedServiced if in Debug, real services if not
 #if DEBUG
-            AddMockedServices(services); //Do i need ref mb here?
+                AddMockedServices(services); //Do i need ref mb here?
 #else
             AddRealServices(services);          
 #endif
+                
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e.Message);
+                Console.WriteLine(e.StackTrace);
+            }     
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            if (env.IsDevelopment())
+            try
             {
-                app.UseDeveloperExceptionPage();
+                if (env.IsDevelopment())
+                {
+                    app.UseDeveloperExceptionPage();
+                }
+
+                app.UseHttpsRedirection();
+
+                app.UseRouting();
+
+                //app.UseAuthorization();
+
+                app.UseEndpoints(endpoints =>
+                {
+                    endpoints.MapControllers();
+                });
             }
-
-            app.UseHttpsRedirection();
-
-            app.UseRouting();
-
-            app.UseAuthorization();
-
-            app.UseEndpoints(endpoints =>
+            catch (Exception e)
             {
-                endpoints.MapControllers();
-            });
+                Console.WriteLine(e.Message);
+                Console.WriteLine(e.StackTrace);
+            }
+         
         }
 
         /// <summary>
