+++ CodeBase/Commons/Domain/Product.cs
@@ -1,9 +1,7 @@
 ﻿
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Linq;
-using System.Text;
 
 namespace Commons.Domain
 {
@@ -274,5 +272,20 @@ namespace Commons.Domain
             return hash;
         }
 
+        public static Product GetMockProduct(string barcode)
+        {
+            List<Product> products = GetMockProducts();
+            return products.Where(x => x.Barcode.Equals(barcode)).FirstOrDefault();
+        }
+
+        public static List<Product> GetMockProducts()
+        {
+            List<Product> products = new List<Product>();
+            products.Add(new Product("fasdfsadfasf", "Apple", "Green Apple", 5.99, 4));
+            products.Add(new Product("fasdfgsdnhrasfassadfasf", "Bannana", "Blue Bananan", 8.99, 5));
+            products.Add(new Product("afafukghkgsfa", "Cat", "Red Care", 566.99, 2));
+
+            return products;
+        }
     }
 }
