+++ CodeBase/Commons/Domain/Transaction.cs
@@ -1,7 +1,6 @@
 ﻿using Commons.Enums;
 using System;
 using System.Collections.Generic;
-using System.Text;
 
 namespace Commons.Domain
 {
@@ -225,6 +224,24 @@ namespace Commons.Domain
             int hash = 179;
             hash = (hash * 19) + ID.GetHashCode();
             return hash;
+        }
+
+        public static Transaction GetMockTransaction(int transactionID)
+        {
+            List<Transaction> transactions = GetMockTransactions();
+            return transactions[transactionID - 1];
+        }
+
+        public static List<Transaction> GetMockTransactions()
+        {
+            List<Transaction> transactions = new List<Transaction>();
+
+            transactions.Add(new Transaction(1, new DateTime(2020, 1, 15), PaymentType.Card, 700));
+            transactions.Add(new Transaction(2, new DateTime(2019, 4, 8), PaymentType.Check, 70));
+            transactions.Add(new Transaction(3, new DateTime(2020, 11, 4), PaymentType.Card, 7200));
+            transactions.Add(new Transaction(4, new DateTime(2028, 3, 7), PaymentType.Card, 800));
+
+            return transactions;
         }
     }
 }
