+++ CodeBase/Commons/Domain/User.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Text;
+using System.Linq;
 
 namespace Commons.Domain
 {
@@ -264,6 +264,50 @@ namespace Commons.Domain
             int hash = 23;
             hash = (hash * 199) + UserName.GetHashCode();
             return hash;
+        }
+
+        public static User GetMockUser(string userName)
+        {
+            List<User> users = GetMockUsers();
+            return users.Where(x => x.UserName == userName).FirstOrDefault();
+        }
+
+        public static List<User> GetMockUsers()
+        {
+            //init
+            List<Basket> mockBasket = Basket.GetMockBaskets();
+            List<User> mockUser = new List<User>();
+
+            //Build
+            mockUser.Add(new User(
+                "Peter",
+                "1334",
+                "xas@google.com",
+                new ObservableCollection<Basket>
+                {
+                    mockBasket[0],
+                    mockBasket[1]
+                }));
+
+            mockUser.Add(new User(
+                "Sven",
+                "133224",
+                "xasffas@google.com",
+                new ObservableCollection<Basket>
+                {
+                    mockBasket[2]
+                }));
+
+            mockUser.Add(new User(
+                "Franz",
+                "21",
+                "Fraaaa@google.com",
+                new ObservableCollection<Basket>
+                {
+                }));
+
+            //return
+            return mockUser;
         }
     }
 }
